Reduce the boilerplate code required to test a component
