+++ syndesis/src/app/App.test.tsx
@@ -5,6 +5,8 @@ import { MemoryRouter } from 'react-router';
 import { App } from './App';
 import i18n from '../i18n';
 
+jest.unmock('./App');
+
 export default describe('App', () => {
   const TestRoute = () => (
     <div data-testid="test-route-container">it renders</div>
